requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1; python_version ~= '3.5'
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cycler==0.10.0
flatbuffers==1.12
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.3; python_version >= '3.6'
google-auth==1.28.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==3.7.3; python_version < '3.8'
joblib==1.0.1; python_version >= '3.6'
keras-applications==1.0.8
keras-preprocessing==1.1.2
keras-vggface==0.6
keras==2.4.3
kiwisolver==1.3.1; python_version >= '3.6'
markdown==3.3.4; python_version >= '3.6'
matplotlib==3.3.4
mtcnn==0.1.0
numpy==1.19.5; python_version >= '3.6'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opencv-python==4.5.1.48; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
pillow==8.1.2; python_version >= '3.6'
protobuf==3.15.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydot==1.4.2
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.7.2; python_version >= '3.6'
scikit-learn==0.24.1; python_version >= '3.6'
scipy==1.5.4; python_version >= '3.6'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==2.4.0
tensorflow==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0; python_version >= '3.5'
typing-extensions==3.7.4.3; python_version < '3.8'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.36.2; python_version >= '3'
wrapt==1.12.1
zipp==3.4.1; python_version >= '3.6'